Add Zod and switch to strict mode

+++ app/api/post/edit/route.ts
@@ -4,10 +4,11 @@ import { PrismaClient } from "@prisma/client";
 import { NextResponse } from "next/server";
 import slugify from "slugify";
 import prisma from "@/prisma";
+import { z } from "zod";
 
 export async function GET(request: Request, response: Response) {
   const { searchParams } = new URL(request.url);
-  const id = parseInt(searchParams.get("id"))!;
+  const id = parseInt(searchParams.get("id")!);
   const session = await getServerSession(OPTIONS);
 
   if (session?.user?.email) {
@@ -28,7 +29,17 @@ export async function POST(request: Request, response: Response) {
 
   if (session?.user?.email) {
     const email = session.user.email;
-    const data = await request.json();
+    const schema = z.object({
+      title: z.string(),
+      content: z.string(),
+      draft: z.boolean(),
+      id: z.string(),
+    });
+
+    const response = schema.safeParse(request.body);
+    if (!response.success) {
+      return NextResponse.json({ error: response.error });
+    }
 
     const user = await prisma.user.findFirst({
       where: { email },
@@ -37,19 +48,21 @@ export async function POST(request: Request, response: Response) {
       },
     });
 
+    const { id, title, content, draft } = response.data;
+
     if (user) {
       const post = await prisma.post.update({
-        where: { id: data.id },
+        where: { id: parseInt(id) },
         data: {
           user: {
             connect: {
               id: user.id,
             },
           },
-          title: data.title,
-          content: data.content,
-          draft: data.draft,
-          slug: slugify(data.title),
+          title,
+          content,
+          draft,
+          slug: slugify(title),
         },
         include: {
           user: true,
@@ -67,12 +80,21 @@ export async function DELETE(request: Request, response: Response) {
     const email = session.user.email;
     const data = await request.json();
 
+    const schema = z.object({
+      id: z.string(),
+    });
+
+    const response = schema.safeParse(data);
+    if (!response.success) {
+      return NextResponse.json({ error: response.error });
+    }
+    const { id } = response.data;
     await prisma.post.delete({
       where: {
         user: {
           email,
         },
-        id: data.id,
+        id: parseInt(id),
       },
     });
     return NextResponse.json({ success: true });
